functional profile page with balance TODO update save to book

+++ model.py
@@ -18,7 +18,7 @@ class User(db.Model):
     password = db.Column(db.String(20), nullable=False)
     email = db.Column(db.String(20), unique=True, nullable=False)
     #payment information default zero
-    balance = db.Column(db.Integer, nullable=False)
+    balance = db.Column(db.Float, nullable=False)
     birthday = db.Column(db.Date, nullable=False)
     #payment_info =
